pkg/hostarch: get page size from unix.Getpagesize
It appears that most usage of PageSize is centralized, or otherwise
relatively trivial, so most if not all the code in the repository should
be working with other page sizes.

This moves the page size into a variable that is initialized at startup,
and makes basic usage of e.g. pkg/abi/linux usable on arm64 systems with
a larger page size.

Additional issues should probably be addressed in-place when they are
discovered.

Updates #8196